LayoutTests/imported/w3c/web-platform-tests/css/cssom-view/interfaces-expected.txt

CONSOLE MESSAGE: line 482: callback not yet supported
CONSOLE MESSAGE: line 482: callback not yet supported
CONSOLE MESSAGE: line 482: callback not yet supported
CONSOLE MESSAGE: line 482: callback not yet supported
CONSOLE MESSAGE: line 482: callback not yet supported
CONSOLE MESSAGE: line 482: callback not yet supported
CONSOLE MESSAGE: line 482: callback not yet supported

CSSOM View IDL tests


PASS Test driver 
PASS Partial interface Window: original interface defined 
PASS Partial interface Document: original interface defined 
PASS Partial interface Element: original interface defined 
PASS Partial interface HTMLElement: original interface defined 
PASS Partial interface HTMLImageElement: original interface defined 
PASS Partial interface Range: original interface defined 
PASS HTMLElement interface: attribute offsetParent 
PASS HTMLElement interface: attribute offsetTop 
PASS HTMLElement interface: attribute offsetLeft 
PASS HTMLElement interface: attribute offsetWidth 
PASS HTMLElement interface: attribute offsetHeight 
PASS HTMLElement interface: document.createElement('div') must inherit property "offsetParent" with the proper type 
PASS HTMLElement interface: document.createElement('div') must inherit property "offsetTop" with the proper type 
PASS HTMLElement interface: document.createElement('div') must inherit property "offsetLeft" with the proper type 
PASS HTMLElement interface: document.createElement('div') must inherit property "offsetWidth" with the proper type 
PASS HTMLElement interface: document.createElement('div') must inherit property "offsetHeight" with the proper type 
PASS Element interface: document.createElement('div') must inherit property "getClientRects()" with the proper type 
PASS Element interface: document.createElement('div') must inherit property "getBoundingClientRect()" with the proper type 
PASS Element interface: document.createElement('div') must inherit property "scrollIntoView([object Object],[object Object])" with the proper type 
PASS Element interface: calling scrollIntoView([object Object],[object Object]) on document.createElement('div') with too few arguments must throw TypeError 
PASS Element interface: document.createElement('div') must inherit property "scroll(ScrollToOptions)" with the proper type 
PASS Element interface: calling scroll(ScrollToOptions) on document.createElement('div') with too few arguments must throw TypeError 
PASS Element interface: document.createElement('div') must inherit property "scroll(unrestricted double, unrestricted double)" with the proper type 
PASS Element interface: calling scroll(unrestricted double, unrestricted double) on document.createElement('div') with too few arguments must throw TypeError 
PASS Element interface: document.createElement('div') must inherit property "scrollTo(ScrollToOptions)" with the proper type 
PASS Element interface: calling scrollTo(ScrollToOptions) on document.createElement('div') with too few arguments must throw TypeError 
PASS Element interface: document.createElement('div') must inherit property "scrollTo(unrestricted double, unrestricted double)" with the proper type 
PASS Element interface: calling scrollTo(unrestricted double, unrestricted double) on document.createElement('div') with too few arguments must throw TypeError 
PASS Element interface: document.createElement('div') must inherit property "scrollBy(ScrollToOptions)" with the proper type 
PASS Element interface: calling scrollBy(ScrollToOptions) on document.createElement('div') with too few arguments must throw TypeError 
PASS Element interface: document.createElement('div') must inherit property "scrollBy(unrestricted double, unrestricted double)" with the proper type 
PASS Element interface: calling scrollBy(unrestricted double, unrestricted double) on document.createElement('div') with too few arguments must throw TypeError 
PASS Element interface: document.createElement('div') must inherit property "scrollTop" with the proper type 
PASS Element interface: document.createElement('div') must inherit property "scrollLeft" with the proper type 
PASS Element interface: document.createElement('div') must inherit property "scrollWidth" with the proper type 
PASS Element interface: document.createElement('div') must inherit property "scrollHeight" with the proper type 
PASS Element interface: document.createElement('div') must inherit property "clientTop" with the proper type 
PASS Element interface: document.createElement('div') must inherit property "clientLeft" with the proper type 
PASS Element interface: document.createElement('div') must inherit property "clientWidth" with the proper type 
PASS Element interface: document.createElement('div') must inherit property "clientHeight" with the proper type 
FAIL Element interface: document.createElement('div') must inherit property "getBoxQuads(BoxQuadOptions)" with the proper type assert_inherits: property "getBoxQuads" not found in prototype chain
FAIL Element interface: calling getBoxQuads(BoxQuadOptions) on document.createElement('div') with too few arguments must throw TypeError assert_inherits: property "getBoxQuads" not found in prototype chain
FAIL Element interface: document.createElement('div') must inherit property "convertQuadFromNode(DOMQuadInit, GeometryNode, ConvertCoordinateOptions)" with the proper type assert_inherits: property "convertQuadFromNode" not found in prototype chain
FAIL Element interface: calling convertQuadFromNode(DOMQuadInit, GeometryNode, ConvertCoordinateOptions) on document.createElement('div') with too few arguments must throw TypeError assert_inherits: property "convertQuadFromNode" not found in prototype chain
FAIL Element interface: document.createElement('div') must inherit property "convertRectFromNode(DOMRectReadOnly, GeometryNode, ConvertCoordinateOptions)" with the proper type assert_inherits: property "convertRectFromNode" not found in prototype chain
FAIL Element interface: calling convertRectFromNode(DOMRectReadOnly, GeometryNode, ConvertCoordinateOptions) on document.createElement('div') with too few arguments must throw TypeError assert_inherits: property "convertRectFromNode" not found in prototype chain
FAIL Element interface: document.createElement('div') must inherit property "convertPointFromNode(DOMPointInit, GeometryNode, ConvertCoordinateOptions)" with the proper type assert_inherits: property "convertPointFromNode" not found in prototype chain
FAIL Element interface: calling convertPointFromNode(DOMPointInit, GeometryNode, ConvertCoordinateOptions) on document.createElement('div') with too few arguments must throw TypeError assert_inherits: property "convertPointFromNode" not found in prototype chain
PASS HTMLImageElement interface: attribute x 
PASS HTMLImageElement interface: attribute y 
PASS HTMLImageElement interface: document.createElement('img') must inherit property "x" with the proper type 
PASS HTMLImageElement interface: document.createElement('img') must inherit property "y" with the proper type 
PASS HTMLElement interface: document.createElement('img') must inherit property "offsetParent" with the proper type 
PASS HTMLElement interface: document.createElement('img') must inherit property "offsetTop" with the proper type 
PASS HTMLElement interface: document.createElement('img') must inherit property "offsetLeft" with the proper type 
PASS HTMLElement interface: document.createElement('img') must inherit property "offsetWidth" with the proper type 
PASS HTMLElement interface: document.createElement('img') must inherit property "offsetHeight" with the proper type 
PASS Element interface: document.createElement('img') must inherit property "getClientRects()" with the proper type 
PASS Element interface: document.createElement('img') must inherit property "getBoundingClientRect()" with the proper type 
PASS Element interface: document.createElement('img') must inherit property "scrollIntoView([object Object],[object Object])" with the proper type 
PASS Element interface: calling scrollIntoView([object Object],[object Object]) on document.createElement('img') with too few arguments must throw TypeError 
PASS Element interface: document.createElement('img') must inherit property "scroll(ScrollToOptions)" with the proper type 
PASS Element interface: calling scroll(ScrollToOptions) on document.createElement('img') with too few arguments must throw TypeError 
PASS Element interface: document.createElement('img') must inherit property "scroll(unrestricted double, unrestricted double)" with the proper type 
PASS Element interface: calling scroll(unrestricted double, unrestricted double) on document.createElement('img') with too few arguments must throw TypeError 
PASS Element interface: document.createElement('img') must inherit property "scrollTo(ScrollToOptions)" with the proper type 
PASS Element interface: calling scrollTo(ScrollToOptions) on document.createElement('img') with too few arguments must throw TypeError 
PASS Element interface: document.createElement('img') must inherit property "scrollTo(unrestricted double, unrestricted double)" with the proper type 
PASS Element interface: calling scrollTo(unrestricted double, unrestricted double) on document.createElement('img') with too few arguments must throw TypeError 
PASS Element interface: document.createElement('img') must inherit property "scrollBy(ScrollToOptions)" with the proper type 
PASS Element interface: calling scrollBy(ScrollToOptions) on document.createElement('img') with too few arguments must throw TypeError 
PASS Element interface: document.createElement('img') must inherit property "scrollBy(unrestricted double, unrestricted double)" with the proper type 
PASS Element interface: calling scrollBy(unrestricted double, unrestricted double) on document.createElement('img') with too few arguments must throw TypeError 
PASS Element interface: document.createElement('img') must inherit property "scrollTop" with the proper type 
PASS Element interface: document.createElement('img') must inherit property "scrollLeft" with the proper type 
PASS Element interface: document.createElement('img') must inherit property "scrollWidth" with the proper type 
PASS Element interface: document.createElement('img') must inherit property "scrollHeight" with the proper type 
PASS Element interface: document.createElement('img') must inherit property "clientTop" with the proper type 
PASS Element interface: document.createElement('img') must inherit property "clientLeft" with the proper type 
PASS Element interface: document.createElement('img') must inherit property "clientWidth" with the proper type 
PASS Element interface: document.createElement('img') must inherit property "clientHeight" with the proper type 
FAIL Element interface: document.createElement('img') must inherit property "getBoxQuads(BoxQuadOptions)" with the proper type assert_inherits: property "getBoxQuads" not found in prototype chain
FAIL Element interface: calling getBoxQuads(BoxQuadOptions) on document.createElement('img') with too few arguments must throw TypeError assert_inherits: property "getBoxQuads" not found in prototype chain
FAIL Element interface: document.createElement('img') must inherit property "convertQuadFromNode(DOMQuadInit, GeometryNode, ConvertCoordinateOptions)" with the proper type assert_inherits: property "convertQuadFromNode" not found in prototype chain
FAIL Element interface: calling convertQuadFromNode(DOMQuadInit, GeometryNode, ConvertCoordinateOptions) on document.createElement('img') with too few arguments must throw TypeError assert_inherits: property "convertQuadFromNode" not found in prototype chain
FAIL Element interface: document.createElement('img') must inherit property "convertRectFromNode(DOMRectReadOnly, GeometryNode, ConvertCoordinateOptions)" with the proper type assert_inherits: property "convertRectFromNode" not found in prototype chain
FAIL Element interface: calling convertRectFromNode(DOMRectReadOnly, GeometryNode, ConvertCoordinateOptions) on document.createElement('img') with too few arguments must throw TypeError assert_inherits: property "convertRectFromNode" not found in prototype chain
FAIL Element interface: document.createElement('img') must inherit property "convertPointFromNode(DOMPointInit, GeometryNode, ConvertCoordinateOptions)" with the proper type assert_inherits: property "convertPointFromNode" not found in prototype chain
FAIL Element interface: calling convertPointFromNode(DOMPointInit, GeometryNode, ConvertCoordinateOptions) on document.createElement('img') with too few arguments must throw TypeError assert_inherits: property "convertPointFromNode" not found in prototype chain
PASS Window interface: operation matchMedia(CSSOMString) 
PASS Window interface: attribute screen 
FAIL Window interface: operation moveTo(long, long) assert_equals: property has wrong .length expected 2 but got 0
FAIL Window interface: operation moveBy(long, long) assert_equals: property has wrong .length expected 2 but got 0
FAIL Window interface: operation resizeTo(long, long) assert_equals: property has wrong .length expected 2 but got 0
FAIL Window interface: operation resizeBy(long, long) assert_equals: property has wrong .length expected 2 but got 0
PASS Window interface: attribute innerWidth 
PASS Window interface: attribute innerHeight 
PASS Window interface: attribute scrollX 
PASS Window interface: attribute pageXOffset 
PASS Window interface: attribute scrollY 
PASS Window interface: attribute pageYOffset 
PASS Window interface: operation scroll(ScrollToOptions) 
PASS Window interface: operation scroll(unrestricted double, unrestricted double) 
PASS Window interface: operation scrollTo(ScrollToOptions) 
PASS Window interface: operation scrollTo(unrestricted double, unrestricted double) 
PASS Window interface: operation scrollBy(ScrollToOptions) 
PASS Window interface: operation scrollBy(unrestricted double, unrestricted double) 
PASS Window interface: attribute screenX 
PASS Window interface: attribute screenY 
PASS Window interface: attribute outerWidth 
PASS Window interface: attribute outerHeight 
PASS Window interface: attribute devicePixelRatio 
PASS Window interface: window must inherit property "matchMedia(CSSOMString)" with the proper type 
PASS Window interface: calling matchMedia(CSSOMString) on window with too few arguments must throw TypeError 
PASS Window interface: window must inherit property "screen" with the proper type 
PASS Window interface: window must inherit property "moveTo(long, long)" with the proper type 
FAIL Window interface: calling moveTo(long, long) on window with too few arguments must throw TypeError assert_throws: Called with 0 arguments function "function () {
            fn.apply(obj, args);
        }" did not throw
PASS Window interface: window must inherit property "moveBy(long, long)" with the proper type 
FAIL Window interface: calling moveBy(long, long) on window with too few arguments must throw TypeError assert_throws: Called with 0 arguments function "function () {
            fn.apply(obj, args);
        }" did not throw
PASS Window interface: window must inherit property "resizeTo(long, long)" with the proper type 
FAIL Window interface: calling resizeTo(long, long) on window with too few arguments must throw TypeError assert_throws: Called with 0 arguments function "function () {
            fn.apply(obj, args);
        }" did not throw
PASS Window interface: window must inherit property "resizeBy(long, long)" with the proper type 
FAIL Window interface: calling resizeBy(long, long) on window with too few arguments must throw TypeError assert_throws: Called with 0 arguments function "function () {
            fn.apply(obj, args);
        }" did not throw
PASS Window interface: window must inherit property "innerWidth" with the proper type 
PASS Window interface: window must inherit property "innerHeight" with the proper type 
PASS Window interface: window must inherit property "scrollX" with the proper type 
PASS Window interface: window must inherit property "pageXOffset" with the proper type 
PASS Window interface: window must inherit property "scrollY" with the proper type 
PASS Window interface: window must inherit property "pageYOffset" with the proper type 
PASS Window interface: window must inherit property "scroll(ScrollToOptions)" with the proper type 
PASS Window interface: calling scroll(ScrollToOptions) on window with too few arguments must throw TypeError 
PASS Window interface: window must inherit property "scroll(unrestricted double, unrestricted double)" with the proper type 
PASS Window interface: calling scroll(unrestricted double, unrestricted double) on window with too few arguments must throw TypeError 
PASS Window interface: window must inherit property "scrollTo(ScrollToOptions)" with the proper type 
PASS Window interface: calling scrollTo(ScrollToOptions) on window with too few arguments must throw TypeError 
PASS Window interface: window must inherit property "scrollTo(unrestricted double, unrestricted double)" with the proper type 
PASS Window interface: calling scrollTo(unrestricted double, unrestricted double) on window with too few arguments must throw TypeError 
PASS Window interface: window must inherit property "scrollBy(ScrollToOptions)" with the proper type 
PASS Window interface: calling scrollBy(ScrollToOptions) on window with too few arguments must throw TypeError 
PASS Window interface: window must inherit property "scrollBy(unrestricted double, unrestricted double)" with the proper type 
PASS Window interface: calling scrollBy(unrestricted double, unrestricted double) on window with too few arguments must throw TypeError 
PASS Window interface: window must inherit property "screenX" with the proper type 
PASS Window interface: window must inherit property "screenY" with the proper type 
PASS Window interface: window must inherit property "outerWidth" with the proper type 
PASS Window interface: window must inherit property "outerHeight" with the proper type 
PASS Window interface: window must inherit property "devicePixelRatio" with the proper type 
PASS WorkerGlobalScope interface: existence and properties of interface object 
PASS DedicatedWorkerGlobalScope interface: existence and properties of interface object 
PASS SharedWorkerGlobalScope interface: existence and properties of interface object 
PASS WorkerNavigator interface: existence and properties of interface object 
PASS WorkerLocation interface: existence and properties of interface object 
PASS Document interface: operation elementFromPoint(double, double) 
PASS Document interface: operation elementsFromPoint(double, double) 
FAIL Document interface: operation caretPositionFromPoint(double, double) assert_own_property: interface prototype object missing non-static operation expected property "caretPositionFromPoint" missing
PASS Document interface: attribute scrollingElement 
FAIL Document interface: operation getBoxQuads(BoxQuadOptions) assert_own_property: interface prototype object missing non-static operation expected property "getBoxQuads" missing
FAIL Document interface: operation convertQuadFromNode(DOMQuadInit, GeometryNode, ConvertCoordinateOptions) assert_own_property: interface prototype object missing non-static operation expected property "convertQuadFromNode" missing
FAIL Document interface: operation convertRectFromNode(DOMRectReadOnly, GeometryNode, ConvertCoordinateOptions) assert_own_property: interface prototype object missing non-static operation expected property "convertRectFromNode" missing
FAIL Document interface: operation convertPointFromNode(DOMPointInit, GeometryNode, ConvertCoordinateOptions) assert_own_property: interface prototype object missing non-static operation expected property "convertPointFromNode" missing
PASS Document interface: document must inherit property "elementFromPoint(double, double)" with the proper type 
PASS Document interface: calling elementFromPoint(double, double) on document with too few arguments must throw TypeError 
PASS Document interface: document must inherit property "elementsFromPoint(double, double)" with the proper type 
PASS Document interface: calling elementsFromPoint(double, double) on document with too few arguments must throw TypeError 
FAIL Document interface: document must inherit property "caretPositionFromPoint(double, double)" with the proper type assert_inherits: property "caretPositionFromPoint" not found in prototype chain
FAIL Document interface: calling caretPositionFromPoint(double, double) on document with too few arguments must throw TypeError assert_inherits: property "caretPositionFromPoint" not found in prototype chain
PASS Document interface: document must inherit property "scrollingElement" with the proper type 
FAIL Document interface: document must inherit property "getBoxQuads(BoxQuadOptions)" with the proper type assert_inherits: property "getBoxQuads" not found in prototype chain
FAIL Document interface: calling getBoxQuads(BoxQuadOptions) on document with too few arguments must throw TypeError assert_inherits: property "getBoxQuads" not found in prototype chain
FAIL Document interface: document must inherit property "convertQuadFromNode(DOMQuadInit, GeometryNode, ConvertCoordinateOptions)" with the proper type assert_inherits: property "convertQuadFromNode" not found in prototype chain
FAIL Document interface: calling convertQuadFromNode(DOMQuadInit, GeometryNode, ConvertCoordinateOptions) on document with too few arguments must throw TypeError assert_inherits: property "convertQuadFromNode" not found in prototype chain
FAIL Document interface: document must inherit property "convertRectFromNode(DOMRectReadOnly, GeometryNode, ConvertCoordinateOptions)" with the proper type assert_inherits: property "convertRectFromNode" not found in prototype chain
FAIL Document interface: calling convertRectFromNode(DOMRectReadOnly, GeometryNode, ConvertCoordinateOptions) on document with too few arguments must throw TypeError assert_inherits: property "convertRectFromNode" not found in prototype chain
FAIL Document interface: document must inherit property "convertPointFromNode(DOMPointInit, GeometryNode, ConvertCoordinateOptions)" with the proper type assert_inherits: property "convertPointFromNode" not found in prototype chain
FAIL Document interface: calling convertPointFromNode(DOMPointInit, GeometryNode, ConvertCoordinateOptions) on document with too few arguments must throw TypeError assert_inherits: property "convertPointFromNode" not found in prototype chain
PASS Element interface: operation getClientRects() 
PASS Element interface: operation getBoundingClientRect() 
PASS Element interface: operation scrollIntoView([object Object],[object Object]) 
PASS Element interface: operation scroll(ScrollToOptions) 
PASS Element interface: operation scroll(unrestricted double, unrestricted double) 
PASS Element interface: operation scrollTo(ScrollToOptions) 
PASS Element interface: operation scrollTo(unrestricted double, unrestricted double) 
PASS Element interface: operation scrollBy(ScrollToOptions) 
PASS Element interface: operation scrollBy(unrestricted double, unrestricted double) 
PASS Element interface: attribute scrollTop 
PASS Element interface: attribute scrollLeft 
PASS Element interface: attribute scrollWidth 
PASS Element interface: attribute scrollHeight 
PASS Element interface: attribute clientTop 
PASS Element interface: attribute clientLeft 
PASS Element interface: attribute clientWidth 
PASS Element interface: attribute clientHeight 
FAIL Element interface: operation getBoxQuads(BoxQuadOptions) assert_own_property: interface prototype object missing non-static operation expected property "getBoxQuads" missing
FAIL Element interface: operation convertQuadFromNode(DOMQuadInit, GeometryNode, ConvertCoordinateOptions) assert_own_property: interface prototype object missing non-static operation expected property "convertQuadFromNode" missing
FAIL Element interface: operation convertRectFromNode(DOMRectReadOnly, GeometryNode, ConvertCoordinateOptions) assert_own_property: interface prototype object missing non-static operation expected property "convertRectFromNode" missing
FAIL Element interface: operation convertPointFromNode(DOMPointInit, GeometryNode, ConvertCoordinateOptions) assert_own_property: interface prototype object missing non-static operation expected property "convertPointFromNode" missing
PASS Element interface: document.createElementNS('x', 'y') must inherit property "getClientRects()" with the proper type 
PASS Element interface: document.createElementNS('x', 'y') must inherit property "getBoundingClientRect()" with the proper type 
PASS Element interface: document.createElementNS('x', 'y') must inherit property "scrollIntoView([object Object],[object Object])" with the proper type 
PASS Element interface: calling scrollIntoView([object Object],[object Object]) on document.createElementNS('x', 'y') with too few arguments must throw TypeError 
PASS Element interface: document.createElementNS('x', 'y') must inherit property "scroll(ScrollToOptions)" with the proper type 
PASS Element interface: calling scroll(ScrollToOptions) on document.createElementNS('x', 'y') with too few arguments must throw TypeError 
PASS Element interface: document.createElementNS('x', 'y') must inherit property "scroll(unrestricted double, unrestricted double)" with the proper type 
PASS Element interface: calling scroll(unrestricted double, unrestricted double) on document.createElementNS('x', 'y') with too few arguments must throw TypeError 
PASS Element interface: document.createElementNS('x', 'y') must inherit property "scrollTo(ScrollToOptions)" with the proper type 
PASS Element interface: calling scrollTo(ScrollToOptions) on document.createElementNS('x', 'y') with too few arguments must throw TypeError 
PASS Element interface: document.createElementNS('x', 'y') must inherit property "scrollTo(unrestricted double, unrestricted double)" with the proper type 
PASS Element interface: calling scrollTo(unrestricted double, unrestricted double) on document.createElementNS('x', 'y') with too few arguments must throw TypeError 
PASS Element interface: document.createElementNS('x', 'y') must inherit property "scrollBy(ScrollToOptions)" with the proper type 
PASS Element interface: calling scrollBy(ScrollToOptions) on document.createElementNS('x', 'y') with too few arguments must throw TypeError 
PASS Element interface: document.createElementNS('x', 'y') must inherit property "scrollBy(unrestricted double, unrestricted double)" with the proper type 
PASS Element interface: calling scrollBy(unrestricted double, unrestricted double) on document.createElementNS('x', 'y') with too few arguments must throw TypeError 
PASS Element interface: document.createElementNS('x', 'y') must inherit property "scrollTop" with the proper type 
PASS Element interface: document.createElementNS('x', 'y') must inherit property "scrollLeft" with the proper type 
PASS Element interface: document.createElementNS('x', 'y') must inherit property "scrollWidth" with the proper type 
PASS Element interface: document.createElementNS('x', 'y') must inherit property "scrollHeight" with the proper type 
PASS Element interface: document.createElementNS('x', 'y') must inherit property "clientTop" with the proper type 
PASS Element interface: document.createElementNS('x', 'y') must inherit property "clientLeft" with the proper type 
PASS Element interface: document.createElementNS('x', 'y') must inherit property "clientWidth" with the proper type 
PASS Element interface: document.createElementNS('x', 'y') must inherit property "clientHeight" with the proper type 
FAIL Element interface: document.createElementNS('x', 'y') must inherit property "getBoxQuads(BoxQuadOptions)" with the proper type assert_inherits: property "getBoxQuads" not found in prototype chain
FAIL Element interface: calling getBoxQuads(BoxQuadOptions) on document.createElementNS('x', 'y') with too few arguments must throw TypeError assert_inherits: property "getBoxQuads" not found in prototype chain
FAIL Element interface: document.createElementNS('x', 'y') must inherit property "convertQuadFromNode(DOMQuadInit, GeometryNode, ConvertCoordinateOptions)" with the proper type assert_inherits: property "convertQuadFromNode" not found in prototype chain
FAIL Element interface: calling convertQuadFromNode(DOMQuadInit, GeometryNode, ConvertCoordinateOptions) on document.createElementNS('x', 'y') with too few arguments must throw TypeError assert_inherits: property "convertQuadFromNode" not found in prototype chain
FAIL Element interface: document.createElementNS('x', 'y') must inherit property "convertRectFromNode(DOMRectReadOnly, GeometryNode, ConvertCoordinateOptions)" with the proper type assert_inherits: property "convertRectFromNode" not found in prototype chain
FAIL Element interface: calling convertRectFromNode(DOMRectReadOnly, GeometryNode, ConvertCoordinateOptions) on document.createElementNS('x', 'y') with too few arguments must throw TypeError assert_inherits: property "convertRectFromNode" not found in prototype chain
FAIL Element interface: document.createElementNS('x', 'y') must inherit property "convertPointFromNode(DOMPointInit, GeometryNode, ConvertCoordinateOptions)" with the proper type assert_inherits: property "convertPointFromNode" not found in prototype chain
FAIL Element interface: calling convertPointFromNode(DOMPointInit, GeometryNode, ConvertCoordinateOptions) on document.createElementNS('x', 'y') with too few arguments must throw TypeError assert_inherits: property "convertPointFromNode" not found in prototype chain
FAIL Text interface: operation getBoxQuads(BoxQuadOptions) assert_own_property: interface prototype object missing non-static operation expected property "getBoxQuads" missing
FAIL Text interface: operation convertQuadFromNode(DOMQuadInit, GeometryNode, ConvertCoordinateOptions) assert_own_property: interface prototype object missing non-static operation expected property "convertQuadFromNode" missing
FAIL Text interface: operation convertRectFromNode(DOMRectReadOnly, GeometryNode, ConvertCoordinateOptions) assert_own_property: interface prototype object missing non-static operation expected property "convertRectFromNode" missing
FAIL Text interface: operation convertPointFromNode(DOMPointInit, GeometryNode, ConvertCoordinateOptions) assert_own_property: interface prototype object missing non-static operation expected property "convertPointFromNode" missing
FAIL Text interface: document.createTextNode('x') must inherit property "getBoxQuads(BoxQuadOptions)" with the proper type assert_inherits: property "getBoxQuads" not found in prototype chain
FAIL Text interface: calling getBoxQuads(BoxQuadOptions) on document.createTextNode('x') with too few arguments must throw TypeError assert_inherits: property "getBoxQuads" not found in prototype chain
FAIL Text interface: document.createTextNode('x') must inherit property "convertQuadFromNode(DOMQuadInit, GeometryNode, ConvertCoordinateOptions)" with the proper type assert_inherits: property "convertQuadFromNode" not found in prototype chain
FAIL Text interface: calling convertQuadFromNode(DOMQuadInit, GeometryNode, ConvertCoordinateOptions) on document.createTextNode('x') with too few arguments must throw TypeError assert_inherits: property "convertQuadFromNode" not found in prototype chain
FAIL Text interface: document.createTextNode('x') must inherit property "convertRectFromNode(DOMRectReadOnly, GeometryNode, ConvertCoordinateOptions)" with the proper type assert_inherits: property "convertRectFromNode" not found in prototype chain
FAIL Text interface: calling convertRectFromNode(DOMRectReadOnly, GeometryNode, ConvertCoordinateOptions) on document.createTextNode('x') with too few arguments must throw TypeError assert_inherits: property "convertRectFromNode" not found in prototype chain
FAIL Text interface: document.createTextNode('x') must inherit property "convertPointFromNode(DOMPointInit, GeometryNode, ConvertCoordinateOptions)" with the proper type assert_inherits: property "convertPointFromNode" not found in prototype chain
FAIL Text interface: calling convertPointFromNode(DOMPointInit, GeometryNode, ConvertCoordinateOptions) on document.createTextNode('x') with too few arguments must throw TypeError assert_inherits: property "convertPointFromNode" not found in prototype chain
PASS Range interface: operation getClientRects() 
PASS Range interface: operation getBoundingClientRect() 
PASS Range interface: new Range() must inherit property "getClientRects()" with the proper type 
PASS Range interface: new Range() must inherit property "getBoundingClientRect()" with the proper type 
FAIL MediaQueryList interface: existence and properties of interface object assert_own_property: self does not have own property "MediaQueryList" expected property "MediaQueryList" missing
FAIL MediaQueryList interface object length assert_own_property: self does not have own property "MediaQueryList" expected property "MediaQueryList" missing
FAIL MediaQueryList interface object name assert_own_property: self does not have own property "MediaQueryList" expected property "MediaQueryList" missing
FAIL MediaQueryList interface: existence and properties of interface prototype object assert_own_property: self does not have own property "MediaQueryList" expected property "MediaQueryList" missing
FAIL MediaQueryList interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "MediaQueryList" expected property "MediaQueryList" missing
FAIL MediaQueryList interface: existence and properties of interface prototype object's @@unscopables property assert_own_property: self does not have own property "MediaQueryList" expected property "MediaQueryList" missing
FAIL MediaQueryList interface: attribute media assert_own_property: self does not have own property "MediaQueryList" expected property "MediaQueryList" missing
FAIL MediaQueryList interface: attribute matches assert_own_property: self does not have own property "MediaQueryList" expected property "MediaQueryList" missing
FAIL MediaQueryList interface: operation addListener(EventListener) assert_own_property: self does not have own property "MediaQueryList" expected property "MediaQueryList" missing
FAIL MediaQueryList interface: operation removeListener(EventListener) assert_own_property: self does not have own property "MediaQueryList" expected property "MediaQueryList" missing
FAIL MediaQueryList interface: attribute onchange assert_own_property: self does not have own property "MediaQueryList" expected property "MediaQueryList" missing
FAIL MediaQueryList must be primary interface of matchMedia('all') assert_own_property: self does not have own property "MediaQueryList" expected property "MediaQueryList" missing
PASS Stringification of matchMedia('all') 
PASS MediaQueryList interface: matchMedia('all') must inherit property "media" with the proper type 
PASS MediaQueryList interface: matchMedia('all') must inherit property "matches" with the proper type 
PASS MediaQueryList interface: matchMedia('all') must inherit property "addListener(EventListener)" with the proper type 
FAIL MediaQueryList interface: calling addListener(EventListener) on matchMedia('all') with too few arguments must throw TypeError assert_throws: Called with 0 arguments function "function () {
            fn.apply(obj, args);
        }" did not throw
PASS MediaQueryList interface: matchMedia('all') must inherit property "removeListener(EventListener)" with the proper type 
FAIL MediaQueryList interface: calling removeListener(EventListener) on matchMedia('all') with too few arguments must throw TypeError assert_throws: Called with 0 arguments function "function () {
            fn.apply(obj, args);
        }" did not throw
FAIL MediaQueryList interface: matchMedia('all') must inherit property "onchange" with the proper type assert_inherits: property "onchange" not found in prototype chain
FAIL MediaQueryListEvent interface: existence and properties of interface object assert_own_property: self does not have own property "MediaQueryListEvent" expected property "MediaQueryListEvent" missing
FAIL MediaQueryListEvent interface object length assert_own_property: self does not have own property "MediaQueryListEvent" expected property "MediaQueryListEvent" missing
FAIL MediaQueryListEvent interface object name assert_own_property: self does not have own property "MediaQueryListEvent" expected property "MediaQueryListEvent" missing
FAIL MediaQueryListEvent interface: existence and properties of interface prototype object assert_own_property: self does not have own property "MediaQueryListEvent" expected property "MediaQueryListEvent" missing
FAIL MediaQueryListEvent interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "MediaQueryListEvent" expected property "MediaQueryListEvent" missing
FAIL MediaQueryListEvent interface: existence and properties of interface prototype object's @@unscopables property assert_own_property: self does not have own property "MediaQueryListEvent" expected property "MediaQueryListEvent" missing
FAIL MediaQueryListEvent interface: attribute media assert_own_property: self does not have own property "MediaQueryListEvent" expected property "MediaQueryListEvent" missing
FAIL MediaQueryListEvent interface: attribute matches assert_own_property: self does not have own property "MediaQueryListEvent" expected property "MediaQueryListEvent" missing
FAIL MediaQueryListEvent must be primary interface of new MediaQueryListEvent('change') assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: MediaQueryListEvent"
FAIL Stringification of new MediaQueryListEvent('change') assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: MediaQueryListEvent"
FAIL MediaQueryListEvent interface: new MediaQueryListEvent('change') must inherit property "media" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: MediaQueryListEvent"
FAIL MediaQueryListEvent interface: new MediaQueryListEvent('change') must inherit property "matches" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: MediaQueryListEvent"
PASS Screen interface: existence and properties of interface object 
PASS Screen interface object length 
PASS Screen interface object name 
PASS Screen interface: existence and properties of interface prototype object 
PASS Screen interface: existence and properties of interface prototype object's "constructor" property 
PASS Screen interface: existence and properties of interface prototype object's @@unscopables property 
PASS Screen interface: attribute availWidth 
PASS Screen interface: attribute availHeight 
PASS Screen interface: attribute width 
PASS Screen interface: attribute height 
PASS Screen interface: attribute colorDepth 
PASS Screen interface: attribute pixelDepth 
PASS Screen must be primary interface of screen 
PASS Stringification of screen 
PASS Screen interface: screen must inherit property "availWidth" with the proper type 
PASS Screen interface: screen must inherit property "availHeight" with the proper type 
PASS Screen interface: screen must inherit property "width" with the proper type 
PASS Screen interface: screen must inherit property "height" with the proper type 
PASS Screen interface: screen must inherit property "colorDepth" with the proper type 
PASS Screen interface: screen must inherit property "pixelDepth" with the proper type 
FAIL CaretPosition interface: existence and properties of interface object assert_own_property: self does not have own property "CaretPosition" expected property "CaretPosition" missing
FAIL CaretPosition interface object length assert_own_property: self does not have own property "CaretPosition" expected property "CaretPosition" missing
FAIL CaretPosition interface object name assert_own_property: self does not have own property "CaretPosition" expected property "CaretPosition" missing
FAIL CaretPosition interface: existence and properties of interface prototype object assert_own_property: self does not have own property "CaretPosition" expected property "CaretPosition" missing
FAIL CaretPosition interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "CaretPosition" expected property "CaretPosition" missing
FAIL CaretPosition interface: existence and properties of interface prototype object's @@unscopables property assert_own_property: self does not have own property "CaretPosition" expected property "CaretPosition" missing
FAIL CaretPosition interface: attribute offsetNode assert_own_property: self does not have own property "CaretPosition" expected property "CaretPosition" missing
FAIL CaretPosition interface: attribute offset assert_own_property: self does not have own property "CaretPosition" expected property "CaretPosition" missing
FAIL CaretPosition interface: operation getClientRect() assert_own_property: self does not have own property "CaretPosition" expected property "CaretPosition" missing
FAIL CaretPosition must be primary interface of document.caretPositionFromPoint(5, 5) assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: document.caretPositionFromPoint is not a function. (In 'document.caretPositionFromPoint(5, 5)', 'document.caretPositionFromPoint' is undefined)"
FAIL Stringification of document.caretPositionFromPoint(5, 5) assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: document.caretPositionFromPoint is not a function. (In 'document.caretPositionFromPoint(5, 5)', 'document.caretPositionFromPoint' is undefined)"
FAIL CaretPosition interface: document.caretPositionFromPoint(5, 5) must inherit property "offsetNode" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: document.caretPositionFromPoint is not a function. (In 'document.caretPositionFromPoint(5, 5)', 'document.caretPositionFromPoint' is undefined)"
FAIL CaretPosition interface: document.caretPositionFromPoint(5, 5) must inherit property "offset" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: document.caretPositionFromPoint is not a function. (In 'document.caretPositionFromPoint(5, 5)', 'document.caretPositionFromPoint' is undefined)"
FAIL CaretPosition interface: document.caretPositionFromPoint(5, 5) must inherit property "getClientRect()" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: document.caretPositionFromPoint is not a function. (In 'document.caretPositionFromPoint(5, 5)', 'document.caretPositionFromPoint' is undefined)"
